requirements.txt
alabaster==0.7.16
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
docutils==0.19
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
Flask==3.1.0
fonttools==4.53.1
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imagesize==1.4.1
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
seaborn-qqplot==0.5.0
Send2Trash==1.8.3
setuptools==70.2.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
Sphinx==6.1.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib.applehelp==1.0.3
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.14.2
style==1.1.6
tabulate==0.9.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.3.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.3.0
urllib3==1.26.14
virtualenv==20.26.3
wcwidth==0.2.6
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whichcraft==0.6.1
wikipedia==1.4.0
xgboost==2.0.3
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
